internal/recomendationModule/requirements.txt
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
DAWG-Python==0.7.2
docopt==0.6.2
grpcio==1.53.0
grpcio-tools==1.53.0
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.2
murmurhash==1.0.9
numpy==1.24.2
packaging==23.0
pathy==0.10.1
preshed==3.0.8
protobuf==4.22.1
psycopg2-binary==2.9.6
pydantic==1.10.7
pymorphy3==1.2.0
pymorphy3-dicts-ru==2.4.417150.4580142
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
ru-core-news-md @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_md-3.5.0/ru_core_news_md-3.5.0-py3-none-any.whl
ru-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.5.0/ru_core_news_sm-3.5.0-py3-none-any.whl
smart-open==6.3.0
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.9
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==1.1.1
